requirements.txt
python-telegram-bot>=20.0.0
openai>=0.27.0
pillow>=9.0.0